pyproject.toml
[project]
name = "ignis"
version = "3.0.0"
description = "Ignis Quant Signal System — trend-following crypto futures signals"
authors = [
    {name = "73nuts"},
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
urls = { Homepage = "https://github.com/73nuts/crypto-signal-bot" }

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ccxt>=4.0.0",
    "python-binance>=1.0.19",
    "requests>=2.32.4",
    "PyYAML>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "schedule>=1.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "Flask>=3.1.1",
    "gunicorn>=21.2.0",
    "pymysql>=1.1.0",
    "DBUtils>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-limiter>=3.5.0",
    "redis>=5.0.0",
    "cashews>=7.0.0",
    "filelock>=3.20.3",
    "pyarrow>=14.0.0",
    "bubus>=1.5.0",
    "aiohttp>=3.13.3",
    "urllib3>=2.6.3",
    "werkzeug>=3.1.5",
    "h11>=0.16.0",
    "structlog>=24.0.0",
    "aiogram>=3.23.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=7.0",
    "pytest-benchmark>=4.0.0",
    "locust>=2.20.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501", "E402", "E712", "F403", "F841"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "-ra", "--strict-markers"]
markers = [
    "unit: unit tests (fast, no external deps)",
    "integration: integration tests (requires DB/API)",
    "e2e: end-to-end tests (requires full environment)",
    "slow: slow tests (>1s)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
minversion = "7.0"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%H:%M:%S"
